I moved the copy for the interface magnetic fields (Bx) outside of the loop over conserved variables.  I don't see any reason to execute the copy 8 times.
I also changed the index for the interface field at the right boundary of
the grid to match that of the other conserved variables.  To see that this
is correct, simply note that the interface fields are in one-to-one
correspondence with the cell average quantities, only they represent
quantities at the right side of the grid cell.  Hence the indexing should
be the same.

One possible complication remains.  Since the interface fields are located
at the right side of the grid cell, the interface field whose index, or
coordinate, matches that of the first ghost cell on the left side of the
grid needs to be updated self consistently.  This makes it inappropriate to
set it through a ghost zone setting routine.  While this is irrelevant
for 1D cartesian calculations, it needs to be kept in mind for 2D and 3D.