The 1D pencil of conserved variables now contains the cell averaged Bx. Hence it contains NVAR elements while the interpolated left and right states as well as the eigen-matrices, eigenvalues, etc. contain NWAVE elements.  The appropriate changes to account for these differing dimensions is in this commit.
In addition, the cell averaged Bx is used directly in calculating the
eigensystems at cell center for the interpolation step.

The magnetic energy density (pressure) calculation now includes the cell
averaged and interface values for Bx assuming that the variation is linear
within a grid cell which is appropriate for cartesian coordinates.  For
multi-dimensional calculations this should be updated to include the
variation of the other field components as well.