For 2D calculations, the reflecting boundaries should make the y-momentum reflecting at y-boundaries.  When I coded up the 2D version of this routine I apparently forgot to make this change and left the x-momentum as switching sign.  This commit fixes this mistake.
This mistake was, however, informative.  Strictly speaking the code should
not have had any funny behavior caused by this error.  The error
essentially placed a slipstream at the boundary, but the gas pressure
was not modified so there should not have been any fluctuations to drive it
out of equilibrium.  What actually happened, however, was the generation
of a strong odd-even coupling error, similar to what one sees when using
SOR to solve Poisson's equation.  For "omega" large, the low frequency
part of the solution is found more quickly, but a high frequency error
will remain (visible in the residuals) with an odd-even type of coupling.

This error might have grown (rather than damped) due to the unsplit scheme,
but it should be revisited again in the future to ensure that this is not
a real problem.  The test problem was the Sod shock tube problem in the
x-direction.